+++ crates/motsu/src/prelude.rs
@@ -1,7 +1,7 @@
 //! Common imports for `motsu` tests.
 pub use crate::{
     context::{
-        Account, Balance, Contract, ContractCall, FromTag, Funding, VMContext,
+        Account, Contract, ContractCall, FromTag, Funding, VMContext,
         DEFAULT_CHAIN_ID,
     },
     revert::ResultExt,