+++ CHANGELOG.md
@@ -10,15 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Account` can be created from string and byte seeds, and can return a signer. #70
+- Implemented `AddressVM` for `Account` and `Contract<T>`. #70
 - Added ability to set the chain ID in tests. #67
 
 ### Changed
 
-- `Funding::balance` is part of a separate `Balance` trait. #70
 
 ### Changed (Breaking)
 
 - `Account` can no longer be created at a predetermined address. #70
+- `Funding::balance` is removed. #70
 - Bump Stylus SDK to v0.8.2. #46
 
 ## [0.6.0] - 2025-03-11